Init: support non-integral zoom in URL parameters

+++ django/applications/catmaid/static/js/init.js
@@ -1028,7 +1028,7 @@ var realInit = function()
 		if ( isNaN( y ) ) y = undefined;
 		if ( values[ "x" ] ) x = parseInt( values[ "x" ] );
 		if ( isNaN( x ) ) x = undefined;
-		if ( values[ "s" ] ) s = parseInt( values[ "s" ] );
+		if ( values[ "s" ] ) s = parseFloat( values[ "s" ] );
         if ( isNaN( s ) ) s = undefined;
         if ( values[ "active_skeleton_id" ] ) init_active_skeleton = parseInt( values[ "active_skeleton_id" ] );
         if ( values[ "active_node_id" ] ) init_active_node_id = parseInt( values[ "active_node_id" ] );
@@ -1066,7 +1066,7 @@ var realInit = function()
 				}
 				sids.push( sid );
 				if ( values[ "s" + i ] )
-					ss.push( parseInt( values[ "s" + i ] ) );
+					ss.push( parseFloat( values[ "s" + i ] ) );
 				else
 					ss.push( NaN );
 				if ( isNaN( sids[ i ] ) || isNaN( ss[ i ] ) )
